bot/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
discord.py==1.6.0
emoji==1.2.0
idna==3.1
multidict==5.1.0
numpy==1.20.2
Pillow==8.2.0
pytz==2021.1
scipy==1.6.2
typing-extensions==3.7.4.3
yarl==1.6.3
